Always allow users to access their own review requests.
In cases where a user assigned a review request to a private group of which they
were not a member, they'd lose access to their own request. This was pretty bad.
I've added an explicit check for ownership to ensure that users always have
access to their own review requests.

Testing done:
Created a review as a non-admin user and assigned it to a private review group.
Saw that I could no longer access it. Applied this change and reloaded, and saw
that I could see it again. New unit test fails before the change and succeeds
after.

Fixes bug 2446.
Reviewed at http://reviews.reviewboard.org/r/3845/