FedoraHosted: Correct the URL for plain files
The URL was not pointing to the raw files, but instead to an HTML representation
of them. With this patch, they correctly point to the raw files.

Testing done:
Created a new FedoraHosted repository and verified that patches could be
applied.

Fixes bug 2897.
Reviewed at http://reviews.reviewboard.org/r/3853/