Try hard to get PYTHONPATH right for subprocesses.
This sets PYTHONPATH in two places, in an attempt to get it right for
all cases where we invoke subprocesses.

Newly generated reviewboard.wsgi files will have PYTHONPATH explicitly
set to include the conf directory.

We also set PYTHONPATH in initialize to that of sys.path, which should
ensure that, at least after the first URL access, we have all the
modules Review Board has access to available to the subprocess.

Fixes bug #2834