Fix issues with the borders/corners around datagrids.
On the datagrids (all but the dashboard), the title header didn't have
rounded corners, making the background stick out just a bit.

On the bottom of the datagrid, the last item did the same thing. The
last entry now should always have rounded corners.

There was a small issue with the padding at the bottom of the box for
datagrids that made the rounding at the bottom look wrong. This was 1px
padding that IE needed. We now target IE specifically for that padding
with a little hack.