Add a --web-server-port argument for rb-site.
This allows the user to configure the port that should be used in the
web server configuration. It defaults to 80, so nothing changes there,
but it should be easier to automate a configuration using a different
port (such as when deploying behind nginx/varnish) now.

This is a more advanced option, and is not exposed during the wizard
flow.