Add a new UserSession model, and redo support for starring.
This introduces a UserSession model, which is a singleton that
represents the user's session. It is intended to replace gUsername, and
to provide the API for handling the watched groups/review requests.

The ReviewGroup model has been rewritten to be on Backbone.js and to
make use of this new API.

The existing ReviewRequest code now uses this API as well.

Unit tests were added to test the basic UserSession singleton logic, and
the individual setStarred functions.