Add a retina version of our logo.
There's now the old standard version of our logo (which has been
replaced with a slightly cleaner version), and a Retina version.

We're using retinajs for this. It gives us a handy .less .at2x() macro,
which will attempt to fetch a @2x version of an image and apply it using
a background-image. This will be used for icons going forward. It also
will look for any <img> tags with a data-at2x= attribute, and fetch
those versions instead.

The one annoying thing is that this will attempt to fetch a @2x of every
<img> we have now. This will be less of a problem as we move to having
retina versions and using spritesheets in more locations, though we may
have to work around it for file attachments.