Move the diff fragment loading into the new DiffViewer model.
This introduces RB.DiffViewer, which replaces and consolidates our old
diff fragment loading code. It cleans up the API a little and removes a
bunch of redundancy. It also simplifies RB.Diff, which was previously
doing too many different things.

This will later be used to store additional state of the diff viewer,
when that's ported over to use Backbone.