Port RB.Diff to Backbone.js.
The resulting RB.Diff is pretty tiny, since it's just a thin wrapper
over BaseResource's functionality now, and since the diff fragment
support has moved into DiffReviewable.

All it's used for now is for uploading new diffs. It provides an
rspNamespace and a function for returning human-readable error strings,
given an error payload.