Convert RB.Screenshot to Backbone.
This is a fairly trivial change to migrate the old screenshot object to
Backbone.js. The only tricky part of this is that I had to add an indirection
step to BaseResourceModel to ensure that the parent object is loaded before
calling url(), because if we don't have the parent links, we won't have a URL
for the object.

Testing done:
Tested setting screenshot captions and deleting screenshots. Ran js-tests.

Reviewed at http://reviews.reviewboard.org/r/3947/