Convert RB.ReviewReply to Backbone.
This is another relatively simple conversion for the data model for ReviewReply
objects to be based on BaseResource.

Testing done:
Replied to an existing review with body_top, body_bottom, and children
CommentReply objects. Published that reply and saw it get saved correctly.
Verified that loading a draft reply worked as expected. Ran js-tests.