Add support for registered GitHub OAuth applications.
Registered GitHub OAuth applications have the added benefits of central
revocation of a key and higher rate limits. While it's not possible for
us to turn this on for every installation out there (since sharing a
client secret is a very bad idea), those who need it can register their
own application entry and set it up with Review Board.

This is a pretty advanced feature, so there's no UI-side configuration
for this. Those who need it are expected to set the
settings.GITHUB_CLIENT_ID and GITHUB_CLIENT_SECRET keys.