Make use of the new support page, and allow support URL customization.
Now on installations, clicking Get Support will take the user to our
support channel, which has some handy links for them. Depending on
whether they have a support contract and are able to message us
directly, they'll have a handy form there as well.

A new Support Settings page has been added. It provides a link to our
support contracts page, and shows them their Install Key (which uniquely
identifies their installation for us). They can provide this to us when
registering software or buying a support contract.

There's also a Custom Support URL field that can be filled in to change
where Get Support goes to. This makes it easy for a company to point to
a wiki or bug tracker or something, so that users will more easily go to
the right place.