Refactor the screenshotThumbnail code to be largely a Backbone view.
The bulk of $.fn.screenshotThumbail has been pulled out into a Backbone
view. The view handles hooking up events to the model and elements, and
performing modification and deletion actions on the screenshot.

$.fn.screenshotThumbnail itself still exists to query the proper
screenshot to populate the view with, and to hook up events to the
editing state for the review request. In time, this will all move into a
new view that owns review request editing.