Fix comment issue bugs.
Changing comment issue states was pretty broken. Some of this was just code that
was never updated when the comments were converted to Backbone. Once I updated
this, I ran into the trouble that existing DiffComment objects were failing to
validate because they didn't have a fileDiffId. This isn't something which is
returned as part of the API responses (for a reason that I haven't figured out
yet), but it's not required when we just want to change the issue state, so I've
changed the validator to only require it if the comment is new.

Testing done: Fixed, dropped, and reopened some issues.
Reviewed at http://reviews.reviewboard.org/r/3998/