Fix a couple issues with resource loading/saving.
BaseResource's URL, when containing an ID, left off the trailing slash.
While it still worked fine, it did result in a redirect, and thus an
extra call to the server for every URL.

Doing a GET had other small but harmless issues. It provided a
contentType and attribute data, even though these weren't wanted in a
GET.