Don't deal with unused files in a parent diff.
When uploading a diff with a parent diff, it's very common to have files
that are in the parent diff that are not in the actual diff going up for
review. This is especially true for large merge branches, or when some
incorrect parent diffs are put up (wrong parameters passed to
post-review/rbt post).

Each unused file gets checked against the repository and then then
thrown away. This is pointless.

We now ignore every file in a parent diff that doesn't have a matching
diff in the actual diff file.