Fix should_update_issue_status to avoid uncaught exceptions.
Every so often, I see a traceback from our servers coming from
BaseCommentResource.should_update_issue_status. The issue here is that
issue_status isn't in the kwargs, but our issue_string_to_status will raise an
exception if the passed-in value is None. This changes it to always return False
in the case that issue_status isn't in kwargs, and fixes up a little bit of
formatting.

Reviewed at http://reviews.reviewboard.org/r/4025/