Fix the PYTHONPATH issue for good (hopefully).
The latest change to set PYTHONPATH to be limited to the location of the
settings_local.py file broke virtualenvs, which no longer set PYTHONPATH
correctly and break rbssh.

PYTHONPATH seems to be a complicated beast when virtualenvs and
pkg_resources comes in, so we're just going to stop dealing with it
here. Instead, we're using a new variable, RBSITE_PYTHONPATH, which will
point to any paths we have within a site that we care about. For now,
this is the conf/ directory. rbssh will specifically look for this and
add the values to sys.path.