Convert ReviewRequest to Backbone.js.
This introduces a new RB.ReviewRequest and a RB.DraftReviewRequest,
both based on BaseResource. These are the last resource classes that
needed to be converted.

RB.ReviewRequest operates on a review request and provides operations
like closing a review request, reopening, checking for updates, and
creating child objects. It has a 'draft' parameter pointing to a
DraftReviewRequest.

DraftReviewRequest handles setting of fields on a review request and
publishing the draft.

All the hacky .cid checks have been removed, meaning we can now rely on
new-style objects for resources.

Some additional work had to be done to transition our editors over to
the new class. These will end up looking nicer when that code is also
transitioned to new classes. For now, they operate on some raw data
(passing API field names to options.data instead of using attribute
names). This is considered temporary.