Move some action handlers to ReviewRequestEditorView.
ReviewRequestEditorView is now responsible for several of the actions
having to do with the draft. Right now, this consists of the various
forms of closing, reopening, and publishing. Down the road, the others
will move here.

Some of this is still less than nice because of our silly redundant IDs
for actions. That will be cleaned up later as that code gets turned into
nice new classes.