Add signals to notify when fetching/checking a file started/finished.
These signals can be used to hook up code to perform an operation when
fetching files or checking the existence of files, or when knowing when
those fetches/checks are complete.

Along with this, the file in diffutils that did time logging of the
fetch operation has been moved into Repository.get_file, in order to
standardize it.