Add support for modifying close descriptions in ReviewRequestEditor.
The close description code has been removed from reviews.js and reworked
for the new review request editor code. Now, each of the three types of
change description inputs have their own fields registered.

Since these all have the same IDs right now (we determine which to put
into the page in the template), the elementID has been changed to a
selector. Along with this, since they all work on the same field in the
review request model, we're using an array of registered fields and not
a map (which we made no use of anyway).