Move more review request editing responsibilities to the model.
While moving things into ReviewRequestEditor*, much of the logic needed
to be in the view until we were more or less complete with the transfer.

Now, draft publishing, field setting and validation live in the model.
The view listens for events and calls into the model for more things.

The ReviewRequest model and associated draft are now instantiated with
the content from the database, meaning we can now work with the raw
content in the model instead of processing data in the DOM.

This last part had an impact on load order in reviews.js, so some things
had to be shifted around.