Fix replying to file attachment comments when the file attachment is removed.
Commenting on a file attachment, then removing the file attachment, then
replying to the comment was broken. This was due to looking up the
associated review_request only using the relationship for active file
attachments. We now check the inactive_review_request relationship too.
This is something that screenshot comments already do, but that fix
never made it to file attachments.

A unit test was added for this.