Add support for self-hosted hosting services.
Now, hosting services an be marked as "self-hosted," which will cause a
"Domain" field to show up for the service above the account entry. Many
of the functions now take a domain.

Accounts now take a domain as well, and we filter registered accounts by
the domain.

Bug trackers have the same sort of setup as well.

GitHub's had a few more changes added to it for cleanup to use the
hosting URL and to not duplicate so much code.

Unit tests were added for all of this.