Improve file attachment comment links.
The file attachment comments in reviews had a link which looks similar to diff
comments. The text of this link is the filename, and the destination was the
download URL.

In the case where the attachment has a review UI, we want the link target to be
the review UI instead of the download URL. This change adds that in (with the
ability for the review UI to customize the link, for example, to link to a
specific page). This also includes some code for the review UI to customize the
text of the link, for example, to list which page the comment is on.

Testing done: Tested this in conjunction with our premium pdf review UI.

Reviewed at http://reviews.reviewboard.org/r/4148/