Create a base Differ class and move some centralized stuff into it.
The two differs now inherit from Differ. This centralizes the
"interesting line" registration, and even the interesting line header
registration.

The filetype header maps have been moved into filetypes.py.

A tiny optimization was made, which was to check the diff compat version
1 (Myers differ) first, and not second. It won't make a huge difference,
but it's silly we didn't do that before.