Prevent the page from jumping when modifying an issue state.
The introduction of the issue summary table had an annoying problem
where the page would scroll slightly after clicking Fixed, Drop, or
Reopen. The reason is that the matching entry in the table would be
added/removed, and that would affect the scroll offset.

We now work around that by updating the scroll offset after modifying
the table to scroll by the appropriate difference. Users won't see it
jump anymore.