Refactor out the remaining comment issue status code.
This introduces a CommentIssueBarView object, which handles everything
having to do with the issue status bars. It can optionally create and
render the element (in the case of the comment dialog), or reuse an
existing element (in the case of the review boxes).

This is a bit different from the original code in $.issueStatus. That
code did too much. Some of the responsibilities have been moved out or
cleaned up. For example, the IssueSummaryTableView now updates itself
when issue statuses change, rather than the issue bar having to update
it.

As part of this, other cleanups were able to be made. The custom
callback registration in CommentIssueManager is gone, and replaced with
standard Backbone events.

CommentDialogView now requires a reviewRequestEditor, which impacted
some tests. Also, ReviewRequestEditor now takes a CommentIssueManager,
in an effort to reduce (and eventually remove) the use of global
variables.

Unit tests were added for this.